requirements.txt
asciitree==0.3.3
attrs==21.2.0
boto3==1.17.90
botocore==1.20.90
Brotli==1.0.9
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloud-files==1.28.1
cloud-volume==4.1.0
colorcet==2.0.6
compressed-segmentation==2.0.1
connected-components-3d==3.2.0
crc32c==2.2
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
deepmerge==0.3.0
deflate==0.3.0
dijkstra3d==1.9.1
dill==0.3.3
DracoPy==0.0.19
edt==2.1.1
et-xmlfile==1.1.0
fasteners==0.16.2
fastremap==1.12.2
fill-voids==2.0.1
fpdf2==2.3.5
fpzip==1.1.4
gevent==21.1.2
google-api-core==1.30.0
google-auth==1.30.2
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
h5py==3.2.1
idna==2.10
imageio==2.9.0
inflection==0.5.1
iniconfig==1.1.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
kimimaro==2.1.1
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.2
multiprocess==0.70.11.1
networkx==2.5.1
numba==0.53.1
numcodecs==0.7.3
numpy==1.20.3
openpyxl==3.0.7
orjson==3.5.3
packaging==20.9
pandas==1.2.4
param==1.10.1
pathos==0.2.7
Pillow==8.2.0
pluggy==0.13.1
posix-ipc==1.0.5
pox==0.2.9
ppft==1.6.6.3
protobuf==3.17.3
psutil==5.4.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.4.8
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pysimdjson==4.0.0
pytest==6.2.4
python-dateutil==2.8.1
python-jsonschema-objects==0.4.0
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
simplejpeg==1.6.1
six==1.16.0
tenacity==7.0.0
threadpoolctl==2.1.0
tifffile==2021.6.6
toml==0.10.2
tqdm==4.61.0
urllib3==1.26.5
zarr==2.8.3
zope.event==4.5.0
zope.interface==5.4.0
zstandard==0.15.2
